refactor: split tools.py — extract reminders, user_profile, python_exec
Phase 1 of tools.py domain split (~5269→~4800 lines in _core).

Moved three most independent domains to their own submodules under
sjtu_agent/agent/tools/:

  _reminders.py    — add/list/remove reminders (99 lines)
  _user_profile.py — get/update user profile (45 lines)
  _python_exec.py  — execute_python tool (46 lines)

Tools are now a package: tools/__init__.py re-exports everything from
_core.py, which imports from the new submodules.  All existing imports
from sjtu_agent.agent.tools continue to work unchanged.

Each submodule exports its own TOOLS_ENTRIES list; _core.py merges
them in the original order.
